requirement.txt
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.15.2
fake-headers==1.0.2
h11==0.14.0
html5lib==1.1
idna==3.10
joblib==1.4.2
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pickle-mixin==1.0.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
threadpoolctl==3.5.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0